fix: update HintTextFontStyle to nullable and improve hint text rendering logic
- Changed HintTextFontStyle property to nullable type for better handling of font styles.
- Simplified hint text rendering logic to use null-coalescing operator for default font style.
- Updated HTML encoding in DataObjectExtensions to use Uri.EscapeDataString for better URI handling.
- Enhanced exception handling in VstsConnectionService for cache read/write operations.
- Added GetTokenKeyString method for consistent token key handling.
- Minor formatting adjustments in WorkItemQueryTileViewModel and ToastNotificationManager.

+++ Source/WindowsRuntime/UI/Notifications/ToastNotificationManager.cs
@@ -53,7 +53,7 @@ namespace Microsoft.Tools.TeamMate.WindowsRuntime.UI.Notifications
             this.ToastOccurred?.Invoke(this, new ToastNotificationManagerEventArgs
             {
                 EventType = ToastNotificationEventType.Dismissed,
-                DismissReason = (ToastDismissalReason)(int)args.Reason
+                DismissReason = (ToastDismissalReason)args.Reason
             });
         }
